+++ demo/app/home/home-page.ts
@@ -1,6 +1,6 @@
 import * as observable from 'tns-core-modules/data/observable';
 import * as pages from 'tns-core-modules/ui/page';
-import {HelloWorldModel} from './main-view-model';
+import {HelloWorldModel} from './home-view-model';
 
 // Event handler for Page 'loaded' event attached in main-page.xml
 export function pageLoaded(args: observable.EventData) {