+++ src/ChromeTabsManagerActivity.ts
@@ -12,6 +12,8 @@ export class ChromeTabsEvent extends Observable {
 export const BROWSER_ACTIVITY_EVENTS = new ChromeTabsEvent();
 
 const KEY_BROWSER_INTENT = 'browserIntent';
+const BROWSER_RESULT_TYPE = 'browserResultType';
+const DEFAULT_RESULT_TYPE = 'dismiss';
 
 /**
  * Manages the custom chrome tabs intent by detecting when it is dismissed by the user and allowing
@@ -20,6 +22,7 @@ const KEY_BROWSER_INTENT = 'browserIntent';
 @JavaProxy('com.proyecto26.inappbrowser.ChromeTabsManagerActivity')
 export class ChromeTabsManagerActivity extends android.app.Activity {
   private mOpened = false;
+  private resultType = null;
 
   constructor() {
     super();
@@ -32,10 +35,14 @@ export class ChromeTabsManagerActivity extends android.app.Activity {
     // This activity gets opened in 2 different ways. If the extra KEY_BROWSER_INTENT is present we
     // start that intent and if it is not it means this activity was started with FLAG_ACTIVITY_CLEAR_TOP
     // in order to close the intent that was started previously so we just close this.
-    if (this.getIntent().hasExtra(KEY_BROWSER_INTENT)) {
+    if (
+      this.getIntent().hasExtra(KEY_BROWSER_INTENT)
+      && (!savedInstanceState || !savedInstanceState.getString(BROWSER_RESULT_TYPE))
+    ) {
       const browserIntent = <Intent>this.getIntent().getParcelableExtra(KEY_BROWSER_INTENT);
       browserIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
       this.startActivity(browserIntent);
+      this.resultType = DEFAULT_RESULT_TYPE;
     } else {
       this.finish();
     }
@@ -50,29 +57,45 @@ export class ChromeTabsManagerActivity extends android.app.Activity {
     if (!this.mOpened) {
       this.mOpened = true;
     } else {
-      BROWSER_ACTIVITY_EVENTS.set('message', 'chrome tabs activity closed');
-      BROWSER_ACTIVITY_EVENTS.set('resultType', 'cancel');
-      BROWSER_ACTIVITY_EVENTS.notify({
-        eventName: 'DismissedEvent',
-        object: BROWSER_ACTIVITY_EVENTS
-      });
+      this.resultType = 'cancel';
       this.finish();
     }
   }
 
   onDestroy(): void {
-    BROWSER_ACTIVITY_EVENTS.set('message', 'chrome tabs activity destroyed');
-    BROWSER_ACTIVITY_EVENTS.set('resultType', 'dismiss');
-    BROWSER_ACTIVITY_EVENTS.notify({
-      eventName: 'DismissedEvent',
-      object: BROWSER_ACTIVITY_EVENTS
-    });
+    if (this.resultType) {
+      switch (this.resultType) {
+        case 'cancel':
+          BROWSER_ACTIVITY_EVENTS.set('message', 'chrome tabs activity closed');
+          BROWSER_ACTIVITY_EVENTS.set('resultType', 'cancel');
+          break;
+        default:
+          BROWSER_ACTIVITY_EVENTS.set('message', 'chrome tabs activity destroyed');
+          BROWSER_ACTIVITY_EVENTS.set('resultType', DEFAULT_RESULT_TYPE);
+          break;
+      }
+      BROWSER_ACTIVITY_EVENTS.notify({
+        eventName: 'DismissedEvent',
+        object: BROWSER_ACTIVITY_EVENTS
+      });
+      this.resultType = null;
+    }
     super.onDestroy();
   }
 
   onNewIntent(intent: Intent): void {
     super.onNewIntent(intent);
     this.setIntent(intent);
+  }
+
+  onRestoreInstanceState(savedInstanceState: Bundle) {
+    super.onRestoreInstanceState(savedInstanceState);
+    this.resultType = savedInstanceState.getString(BROWSER_RESULT_TYPE);
+  }
+
+  onSaveInstanceState(savedInstanceState: Bundle) {
+    savedInstanceState.putString(BROWSER_RESULT_TYPE, DEFAULT_RESULT_TYPE);
+    super.onSaveInstanceState(savedInstanceState);
   }
 }